+++ orm-patterns-instances/active-jdbc/src/main/java/org/orm/patterns/instances/active/jdbc/service/PhoneServiceExecution.java
@@ -77,7 +77,6 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
      */
     private void createExecution() {
         try {
-
             PersonDetail createPersonRequest = PersonDetail.builder()
                     .firstName(Optional.ofNullable("Person"))
                     .lastName(Optional.ofNullable("Personson"))
@@ -120,7 +119,6 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
             PersonDetail createPersonResponse = this.personService.create(createPersonRequest);
             log.info(ANSI_CYAN_BACKGROUND + "Receive the created Person data - {}" + ANSI_RESET, createPersonResponse);
 
-
             PhoneDetail createPhoneRequest = PhoneDetail.builder()
                     .personId(Optional.ofNullable(createPersonResponse.getId().orElse(1L)))
                     .number(Optional.ofNullable("+7 (999) 999-9999"))
@@ -154,6 +152,9 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
         }
     }
 
+    /**
+     * Delete execution.
+     */
     private void deleteExecution() {
         try {
             Collection<PhoneDetail> findPhoneResponse = new ArrayList<>();