+++ orm-patterns-instances/hibernate-jpa/src/main/java/org/orm/patterns/instances/hibernate/jpa/service/PhoneService.java
@@ -12,6 +12,13 @@ import org.orm.patterns.instances.hibernate.jpa.repository.PersonRepository;
 import org.orm.patterns.instances.hibernate.jpa.repository.PhoneRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
+import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
+import static org.hibernate.validator.internal.util.Contracts.assertTrue;
+
 /**
  * The type Phone service.
  */
@@ -50,8 +57,9 @@ public class PhoneService {
             throws CustomModelMapper.MappingException {
         log.info("Получили запрос на создание сущности - {}", createPhoneRequest);
         Phone phone = this.phoneModelMapper.map(createPhoneRequest, new Phone());
-        Person person = this.personRepository.getOne(createPhoneRequest.getPersonId().orElse(null));
-        log.info("Получили сущность - {}");
+        Person person = this.personRepository.findById(createPhoneRequest.getPersonId().orElse(null))
+                .orElse(null);
+        log.info("Получили сущность - {}", person);
         phone.setPerson(person);
         log.info("Создали сущность - {}", phone);
 
@@ -64,5 +72,84 @@ public class PhoneService {
         log.info("Возвращаем ответ - {}", createPhoneResponse);
 
         return createPhoneResponse;
+    }
+
+    /**
+     * Update phone detail.
+     *
+     * @param idPhone            the uid phone
+     * @param updatePhoneRequest the update phone request
+     * @return the phone detail
+     * @throws CustomModelMapper.MappingException the mapping exception
+     */
+    public PhoneDetail update(Long idPhone, PhoneDetail updatePhoneRequest)
+            throws CustomModelMapper.MappingException {
+        log.info(
+                "Получили запрос на обновлении сущности - {}\n с идентификатором - {}",
+                updatePhoneRequest,
+                idPhone);
+
+        Phone phone = this.phoneRepository.findById(idPhone).orElse(null);
+        assertNotNull(
+                phone,
+                String.format("Сущности с таким идентификатором - {} не существует", idPhone)
+        );
+        log.info("Получили сущность - {}", phone);
+
+        this.phoneModelMapper.map(updatePhoneRequest, phone);
+        this.phoneRepository.save(phone);
+
+        PhoneDetail updatePhoneResponse = this.phoneModelMapper.map(
+                phone,
+                new PhoneDetail()
+        );
+        log.info("Возвращаем ответ - {}", updatePhoneResponse);
+
+        return updatePhoneResponse;
+    }
+
+    /**
+     * Find list.
+     *
+     * @return the list
+     * @throws CustomModelMapper.MappingException the mapping exception
+     */
+    public List<PhoneDetail> find() throws CustomModelMapper.MappingException {
+        List<Phone> phones = this.phoneRepository.findAll();
+        List<PhoneDetail> findPhoneResponse = this.phoneModelMapper.map(
+                phones,
+                PhoneDetail.class
+        );
+        log.info("Возвращаем ответ - {}", findPhoneResponse);
+
+        return findPhoneResponse;
+    }
+
+    /**
+     * Delete phone detail.
+     *
+     * @param idPhone the uid phone
+     * @return the phone detail
+     * @throws CustomModelMapper.MappingException the mapping exception
+     */
+    public PhoneDetail delete(Long idPhone) {
+        log.info("Получили запрос на удаление сущности с идентификатором - {}", idPhone);
+        assertTrue(
+                this.phoneRepository.existsById(idPhone),
+                String.format("Сущности с таким идентификатором - {} не существует", idPhone)
+        );
+
+        this.phoneRepository.deleteById(idPhone);
+        assertTrue(
+                !this.phoneRepository.existsById(idPhone),
+                String.format("Не получилось удалить сущность с идентификатором - {}", idPhone)
+        );
+
+        PhoneDetail deletePhoneResponse = PhoneDetail.builder()
+                .id(Optional.ofNullable(idPhone))
+                .build();
+        log.info("Возвращаем ответ - {}", deletePhoneResponse);
+
+        return deletePhoneResponse;
     }
 }