+++ orm-patterns-instances/hibernate-jpa/src/main/java/org/orm/patterns/instances/hibernate/jpa/service/PhoneServiceExecution.java
@@ -11,10 +11,11 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Optional;
 
-import static org.orm.patterns.instances.commons.util.OutColorsUtils.ANSI_CYAN_BACKGROUND;
-import static org.orm.patterns.instances.commons.util.OutColorsUtils.ANSI_RESET;
+import static org.orm.patterns.instances.commons.util.OutColorsUtils.*;
 
 /**
  * The type Person service execution.
@@ -48,6 +49,18 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
 
         //create new
         this.createExecution();
+        this.findExecution();
+
+        //update exists
+        this.updateExecution();
+        this.findExecution();
+
+        //get all
+        this.findExecution();
+
+        //delete exists
+        this.deleteExecution();
+        this.findExecution();
 
         return;
     }
@@ -73,6 +86,88 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
             log.info(ANSI_CYAN_BACKGROUND + "Receive the created Phone data - {}" + ANSI_RESET, createPhoneResponse);
         } catch (CustomModelMapper.MappingException e) {
             log.debug("Couldn't create the new Person, because - {}", e.getMessage());
+        }
+    }
+
+    /**
+     * Update execution.
+     */
+    private void updateExecution() {
+        try {
+            PersonDetail createPersonRequest = PersonDetail.builder()
+                    .firstName(Optional.ofNullable("Person"))
+                    .lastName(Optional.ofNullable("Personson"))
+                    .age(Optional.ofNullable(16))
+                    .build();
+            PersonDetail createPersonResponse = this.personService.create(createPersonRequest);
+            log.info(ANSI_CYAN_BACKGROUND + "Receive the created Person data - {}" + ANSI_RESET, createPersonResponse);
+
+            PhoneDetail createPhoneRequest = PhoneDetail.builder()
+                    .personId(Optional.ofNullable(createPersonResponse.getId().orElse(1L)))
+                    .number(Optional.ofNullable("+7 (999) 999-9999"))
+                    .build();
+            PhoneDetail createPhoneResponse = this.phoneService.create(createPhoneRequest);
+            Long updatePhoneIdRequest = createPhoneResponse.getId().get();
+
+            PhoneDetail updatePhoneRequest = PhoneDetail.builder()
+                    .personId(Optional.ofNullable(createPersonResponse.getId().orElse(1L)))
+                    .number(Optional.ofNullable("+7 (111) 111-1111"))
+                    .build();
+            PhoneDetail updatePhoneResponse = this.phoneService.update(updatePhoneIdRequest, updatePhoneRequest);
+            log.info(ANSI_CYAN_BACKGROUND + "Receive the updated Phone data - {}" + ANSI_RESET, updatePhoneResponse);
+        } catch (CustomModelMapper.MappingException e) {
+            log.debug("Couldn't create the new Phone, because - {}", e.getMessage());
+        }
+    }
+
+    /**
+     * Find execution.
+     */
+    private void findExecution() {
+        try {
+            Collection<PhoneDetail> findPhoneResponse = this.phoneService.find();
+            log.info(
+                    ANSI_CYAN_BACKGROUND + "Receive the collection of Phones data - {}" + ANSI_RESET,
+                    findPhoneResponse
+            );
+        } catch (CustomModelMapper.MappingException e) {
+            log.debug("Couldn't find any Phone, because - {}", e.getMessage());
+        }
+    }
+
+    /**
+     * Delete execution.
+     */
+    private void deleteExecution() {
+        try {
+            Collection<PhoneDetail> findPhoneResponse = new ArrayList<>();
+            if ((findPhoneResponse = this.phoneService.find()).isEmpty()) {
+                PersonDetail createPersonRequest = PersonDetail.builder()
+                        .firstName(Optional.ofNullable("Person"))
+                        .lastName(Optional.ofNullable("Personson"))
+                        .age(Optional.ofNullable(16))
+                        .build();
+                PersonDetail createPersonResponse = this.personService.create(createPersonRequest);
+                log.info(ANSI_CYAN_BACKGROUND + "Receive the created Person data - {}" + ANSI_RESET, createPersonResponse);
+
+                PhoneDetail createPhoneRequest = PhoneDetail.builder()
+                        .personId(Optional.ofNullable(createPersonResponse.getId().orElse(1L)))
+                        .number(Optional.ofNullable("+7 (999) 999-9999"))
+                        .build();
+                findPhoneResponse.add(this.phoneService.create(createPhoneRequest));
+            }
+
+            this.phoneService.find().stream()
+                    .forEach(phone -> {
+                        Long deletePhoneIdRequest = phone.getId().get();
+                        PhoneDetail deletePhoneResponse = this.phoneService.delete(deletePhoneIdRequest);
+                        log.info(
+                                ANSI_CYAN_BACKGROUND + "Receive the deleted Phone data - {}" + ANSI_RESET,
+                                deletePhoneResponse
+                        );
+                    });
+        } catch (Exception e) {
+            log.debug("Couldn't delete the Phone, because - {}", e.getMessage());
         }
     }
 }